requirements.txt
selenium==4.18.1
requests==2.31.0
beautifulsoup4==4.12.3
pyyaml==6.0.1
python-dotenv==1.0.1
ollama==0.1.6
pillow==10.2.0
urllib3==2.2.1
webdriver-manager==4.0.1
tqdm==4.66.2 